+++ src/components/TodoList.js
@@ -1,5 +1,5 @@
-import { setItem } from "./storage.js";
-import validation from "./util/validation.js";
+import { setItem } from "../utils/storage.js";
+import validation from "../utils/validation.js";
 
 // params.$target - 해당 컴포넌트가 추가될 DOM 앨리먼트
 // params.initialState - 해당 컴포넌트의 초기 상태
@@ -27,8 +27,8 @@ export default function TodoList({ $target, initialState, updateCount }) {
           .map(
             ({ text, isCompleted }, index) => `
               <li data-index=${index} class="todoList ${
-              isCompleted ? "completed" : ""
-            }">
+                isCompleted ? "completed" : ""
+              }">
                 ${text}
                 <button class="deleteBtn">삭제</button>
               </li>
